feat(gradient): allow out-of-range stop positions
Move renormalize_stops_to_unit_range from convert.rs to background.rs
(near the resolve_gradient_stops call site) and wire it in: drop the
log::warn bail for out-of-range fractions, instead synthesize boundary
stops via linear interpolation per CSS Images 3 §3.5.1.

Also drops the now-unused gradient_kind argument and updates the two
existing tests (out_of_range_length_returns_none -> _renormalized,
negative_fraction_returns_none -> _renormalized) to verify the new
behavior instead of asserting None. The renormalize helper signature
is simplified to Vec<(f32, [u8; 4])> -> Vec<(f32, [u8; 4])> since
call sites build krilla::paint::Stop directly.

For fulgur-n3zk.
