+++ crates/fulgur/src/convert.rs
@@ -3201,110 +3201,6 @@ fn resolve_linear_gradient(
     ))
 }
 
-/// 範囲外 stop を CSS Images 3 §3.5.1 準拠で [0, 1] 内表現に変換する。
-///
-/// 入力: monotonically non-decreasing position の (pos, rgba) ベクタ。
-/// pos は ℝ で、範囲外 (-0.5 や 1.5 など) も許容する。
-///
-/// 出力: krilla `NormalizedF32` 制約を満たす GradientStop ベクタ
-/// (position は `Fraction(f)` で `f ∈ [0, 1]`)。
-///
-/// アルゴリズム:
-/// - fast path: 全 stop が [0, 1] 内なら無変換
-/// - 範囲外を含む場合: `color_at(0)` / `color_at(1)` を隣接 stop 線形補間
-///   (または pad 前方/後方) で合成し、(0, 1) 内既存 stop と組み合わせる
-///
-/// この helper は Task 1 (fulgur-n3zk) の TDD 単独追加。Task 2 で
-/// `resolve_color_stops` 経路から呼び出されるまでは call site なしのため
-/// `#[allow(dead_code)]` を付ける。
-#[allow(dead_code)]
-fn renormalize_stops_to_unit_range(
-    stops: Vec<(f32, [u8; 4])>,
-) -> Vec<crate::pageable::GradientStop> {
-    use crate::pageable::{GradientStop, GradientStopPosition};
-
-    debug_assert!(stops.len() >= 2, "caller guaranteed len >= 2");
-
-    if stops.iter().all(|(p, _)| (0.0..=1.0).contains(p)) {
-        return stops
-            .into_iter()
-            .map(|(p, rgba)| GradientStop {
-                position: GradientStopPosition::Fraction(p),
-                rgba,
-            })
-            .collect();
-    }
-
-    let last_idx = stops.len() - 1;
-    let color_at = |t: f32| -> [u8; 4] {
-        if t <= stops[0].0 {
-            return stops[0].1;
-        }
-        if t >= stops[last_idx].0 {
-            return stops[last_idx].1;
-        }
-        for w in stops.windows(2) {
-            let (p0, c0) = w[0];
-            let (p1, c1) = w[1];
-            if p0 <= t && t <= p1 {
-                let span = p1 - p0;
-                if span <= 0.0 {
-                    return c1;
-                }
-                let alpha = (t - p0) / span;
-                return [
-                    lerp_u8(c0[0], c1[0], alpha),
-                    lerp_u8(c0[1], c1[1], alpha),
-                    lerp_u8(c0[2], c1[2], alpha),
-                    lerp_u8(c0[3], c1[3], alpha),
-                ];
-            }
-        }
-        unreachable!("stops are monotonic and t is in [stops[0].0, stops[last].0]")
-    };
-
-    // Skip boundary synthesis when an in-range stop already sits exactly on
-    // that boundary — otherwise the synthesized stop would duplicate the
-    // existing one (same offset, same color via linear interpolation).
-    let has_stop_at_zero = stops.iter().any(|(p, _)| *p == 0.0);
-    let has_stop_at_one = stops.iter().any(|(p, _)| *p == 1.0);
-
-    let mut result = Vec::with_capacity(stops.len() + 2);
-
-    if stops[0].0 != 0.0 && !has_stop_at_zero {
-        result.push(GradientStop {
-            position: GradientStopPosition::Fraction(0.0),
-            rgba: color_at(0.0),
-        });
-    }
-
-    for (p, rgba) in stops.iter() {
-        if (0.0..=1.0).contains(p) {
-            result.push(GradientStop {
-                position: GradientStopPosition::Fraction(*p),
-                rgba: *rgba,
-            });
-        }
-    }
-
-    if stops[last_idx].0 != 1.0 && !has_stop_at_one {
-        result.push(GradientStop {
-            position: GradientStopPosition::Fraction(1.0),
-            rgba: color_at(1.0),
-        });
-    }
-
-    result
-}
-
-#[inline]
-#[allow(dead_code)]
-fn lerp_u8(a: u8, b: u8, alpha: f32) -> u8 {
-    let av = a as f32;
-    let bv = b as f32;
-    (av + (bv - av) * alpha).round().clamp(0.0, 255.0) as u8
-}
-
 /// CSS gradient items から GradientStop ベクタを解決する。linear / radial 共通。
 ///
 /// position は `GradientStopPosition` で保持され (Auto / Fraction / LengthPx)、
@@ -5715,107 +5611,5 @@ mod inline_box_extraction_tests {
             expected,
             delta
         );
-    }
-}
-
-#[cfg(test)]
-mod gradient_renormalize_tests {
-    use super::*;
-    use crate::pageable::{GradientStop, GradientStopPosition};
-
-    fn s(offset: f32, r: u8, g: u8, b: u8) -> (f32, [u8; 4]) {
-        (offset, [r, g, b, 255])
-    }
-
-    fn fraction_offset(stop: &GradientStop, idx: usize) -> f32 {
-        match stop.position {
-            GradientStopPosition::Fraction(f) => f,
-            other => panic!("stop[{idx}].position: expected Fraction, got {other:?}"),
-        }
-    }
-
-    fn expect(stops: &[GradientStop], expected: &[(f32, [u8; 4])]) {
-        assert_eq!(stops.len(), expected.len(), "stop count mismatch");
-        for (i, (got, exp)) in stops.iter().zip(expected).enumerate() {
-            let got_offset = fraction_offset(got, i);
-            assert!(
-                (got_offset - exp.0).abs() < 1e-5,
-                "stop[{i}].offset: got {got_offset} expected {}",
-                exp.0
-            );
-            assert_eq!(got.rgba, exp.1, "stop[{i}].rgba");
-        }
-    }
-
-    #[test]
-    fn no_op_when_all_in_range() {
-        let stops = vec![s(0.0, 255, 0, 0), s(1.0, 0, 0, 255)];
-        let result = renormalize_stops_to_unit_range(stops);
-        expect(&result, &[(0.0, [255, 0, 0, 255]), (1.0, [0, 0, 255, 255])]);
-    }
-
-    #[test]
-    fn synthesize_left_endpoint() {
-        // red at -50%, blue at 100%: at offset 0, t = 0.5 / 1.5 = 1/3
-        // r = 255 + 1/3 * (0 - 255) = 170
-        // g = 0
-        // b = 0   + 1/3 * (255 - 0) = 85
-        let stops = vec![s(-0.5, 255, 0, 0), s(1.0, 0, 0, 255)];
-        let result = renormalize_stops_to_unit_range(stops);
-        expect(
-            &result,
-            &[(0.0, [170, 0, 85, 255]), (1.0, [0, 0, 255, 255])],
-        );
-    }
-
-    #[test]
-    fn synthesize_right_endpoint() {
-        // red at 50%, blue at 200%: at offset 1, t = 0.5 / 1.5 = 1/3
-        let stops = vec![s(0.5, 255, 0, 0), s(2.0, 0, 0, 255)];
-        let result = renormalize_stops_to_unit_range(stops);
-        expect(
-            &result,
-            &[
-                (0.0, [255, 0, 0, 255]),
-                (0.5, [255, 0, 0, 255]),
-                (1.0, [170, 0, 85, 255]),
-            ],
-        );
-    }
-
-    #[test]
-    fn all_below_zero_pads_with_last_color() {
-        // stops at -50%, -25%: both out of range; pad after last (blue)
-        let stops = vec![s(-0.5, 255, 0, 0), s(-0.25, 0, 0, 255)];
-        let result = renormalize_stops_to_unit_range(stops);
-        expect(&result, &[(0.0, [0, 0, 255, 255]), (1.0, [0, 0, 255, 255])]);
-    }
-
-    #[test]
-    fn all_above_one_pads_with_first_color() {
-        // stops at 150%, 200%: both out of range; pad before first (red)
-        let stops = vec![s(1.5, 255, 0, 0), s(2.0, 0, 0, 255)];
-        let result = renormalize_stops_to_unit_range(stops);
-        expect(&result, &[(0.0, [255, 0, 0, 255]), (1.0, [255, 0, 0, 255])]);
-    }
-
-    #[test]
-    fn boundary_stop_at_zero_no_left_synthesis() {
-        // red at -50%, blue at 0%, green at 100%
-        // 0 はちょうど blue の位置なので合成不要
-        let stops = vec![s(-0.5, 255, 0, 0), s(0.0, 0, 0, 255), s(1.0, 0, 255, 0)];
-        let result = renormalize_stops_to_unit_range(stops);
-        expect(&result, &[(0.0, [0, 0, 255, 255]), (1.0, [0, 255, 0, 255])]);
-    }
-
-    #[test]
-    fn alpha_channel_is_interpolated() {
-        // red(alpha=0) at -50%, blue(alpha=255) at 100%
-        // at offset 0, t = 1/3 → alpha = 0 + 1/3 * 255 ≈ 85
-        let stops = vec![(-0.5_f32, [255, 0, 0, 0]), (1.0_f32, [0, 0, 255, 255])];
-        let result = renormalize_stops_to_unit_range(stops);
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0].rgba[3], 85, "alpha at offset 0");
-        assert_eq!(result[1].rgba[3], 255, "alpha at offset 1");
     }
 }
